pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "himu"
version = "1.0.0"
description = "HiMu: Hierarchical Multimodal Frame Selection for Long Video QA"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Dan Baras"},
]
keywords = ["video-qa", "frame-selection", "multimodal", "lvlm", "video-understanding"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "pydantic>=2.0",
    "openai>=1.0",
    "opencv-python>=4.8",
]

[project.optional-dependencies]
clip = [
    "open-clip-torch>=2.24",
    "torch>=2.1",
    "pillow",
]
yolo = [
    "ultralytics>=8.1",
    "torch>=2.1",
]
ocr = [
    "python-doctr[torch]>=0.8",
    "torch>=2.1",
]
asr = [
    "faster-whisper>=1.0",
    "soundfile",
    "librosa>=0.10",
    "sentence-transformers>=2.2",
]
clap = [
    "transformers>=4.35",
    "torch>=2.1",
    "librosa>=0.10",
]
all = ["himu[clip,yolo,ocr,asr,clap]"]
dev = [
    "pytest>=7.0",
    "pytest-cov",
]

[project.urls]
Homepage = "https://github.com/danbaras/himu"

[tool.setuptools.packages.find]
include = ["himu*"]
